[minor](refactor) rename parquet predicate
